Update README.md
Your company is trying to decide on the most cost effective approach to build a new environment to support a new business initiative.   Do they build onsite or use a cloud provider?

You will conduct a Cost Estimate and Cost Analysis that will compare the cost estimate of two separate infrastructure with the same infrastructure requirements:  AWS Infrastructure (cloud provider) and Traditional Infrastructure (onsite).   You are required to provide the Cost Estimate information and answer the Cost Analysis questions.

Infrastructure requirements:
Server Farm 1: 8 Servers with the following specifications each: Dual 4 CPU Core Processors, 64GB RAM, Linux, 250GB SSD 
Server Farm 2: 12 Servers with the following specifications each: Dual 8 CPU Core Processors, 256GB RAM, Windows, 1TB SSD
Deliverables
Create a Word document that has the Cost Estimate and Cost Analysis.  APA is required for this assignment. This assignment's reference requirement is to cite minimum of two separate references that do not relate to your cost estimate.  One reference is your textbook and the other reference is an external source (i.e., journal article, another textbook, not a blog, not Wikipedia).  

Cost Estimate
AWS Infrastructure Estimate: Use the AWS pricing calculator AWS Calculator Links to an external site.to calculate the AWS costs for the above infrastructure requirements.  Include a calculator screenshot of the estimate for all services.
Traditional Infrastructure Estimate: Use any website to calculate the traditional infrastructure cost for the above infrastructure requirements. Include a picture and price for each device type (each server farm will require a different model of server).  While the hardware will not be an exact match to AWS, make it as close as you can!  Do try to make sure you have enterprise-level servers, which will include a RAID controller, multiple HDDs and will be rack mountable.
Cost Comparison: Create tables based on the one-year cost of AWS:
A table of the resources needed for each environment. Include the following columns: Description of resource type (server type), quantity, and total cost. 
A table with the line item price for AWS and and a table with the line item price of physical infrastructure so the two can be compared and contrasted.
A table which compares the total price of AWS and Physical Infrastructure.
Cost Analysis
Write 400 - 550 words answering the following questions based on your cost estimate:

Compare the cost of building an environment in AWS to a Traditional infrastructure. Which option costs more and why?
Total Cost of Ownership is more than the initial price.  Describe two costs of traditional infrastructure that are not accounted for in your estimate.
Describe two cost savings of the cloud that are not accounted for in your estimate. Do not use the same items from #2!
Hints
When calculating AWS costs, consider reserved instances! This can be done in the AWS calculator.
There is not an expectation of perfect numbers; the point of this assignment is to get you to perform your own research!  Any path that fulfills the requirements of this lab is fine.
You may not find an apples to apples comparison between the two options... they are very different than one another so this is to be expected!  Do your best to align the two options.